refactor fx tweakpane into components
