+++ src/components/Params/Pane.module.scss
@@ -1,0 +1,4 @@
+.pane {
+  margin-bottom: 10px;
+  margin-top: 10px;
+}
