+++ src/components/Params/tweakpane/plugins/StringInputPlugin/index.ts
@@ -7,14 +7,14 @@ import {
   InputBindingPlugin,
   ParamsParsers,
   parseParams,
+  ValueMap,
+  Constraint,
 } from "@tweakpane/core"
-import { PluginController } from "./controller"
+import { FxStringInputController } from "./controller"
 
-export interface PluginInputParams extends BaseInputParams {
-  max?: number
-  min?: number
-  step?: number
-  view: "dots"
+export interface IFxStringInputPluginParams extends BaseInputParams {
+  view: "string"
+  maxLength?: number
 }
 
 // NOTE: You can see JSDoc comments of `InputBindingPlugin` for details about each property
@@ -24,69 +24,50 @@ export interface PluginInputParams extends BaseInputParams {
 // - converts `Ex` into `In` and holds it
 // - P is the type of the parsed parameters
 //
-export const TemplateInputPlugin: InputBindingPlugin<
-  number,
-  number,
-  PluginInputParams
+export const FxStringInputPlugin: InputBindingPlugin<
+  string,
+  string,
+  IFxStringInputPluginParams
 > = {
-  id: "input-template",
+  id: "fx-string-input-template",
 
   // type: The plugin type.
   // - 'input': Input binding
   // - 'monitor': Monitor binding
   type: "input",
 
-  // This plugin template injects a compiled CSS by @rollup/plugin-replace
-  // See rollup.config.js for details
-  css: "",
-
-  accept(exValue: unknown, params: Record<string, unknown>) {
-    if (typeof exValue !== "number") {
-      // Return null to deny the user input
+  accept(value: unknown, params: Record<string, unknown>) {
+    if (typeof value !== "string") {
       return null
     }
-
     // Parse parameters object
     const p = ParamsParsers
-    const result = parseParams<PluginInputParams>(params, {
+    const result = parseParams<IFxStringInputPluginParams>(params, {
       // `view` option may be useful to provide a custom control for primitive values
-      view: p.required.constant("dots"),
-
-      max: p.optional.number,
-      min: p.optional.number,
-      step: p.optional.number,
+      view: p.required.constant("string"),
+      maxLength: p.optional.number,
     })
     if (!result) {
       return null
     }
-
     // Return a typed value and params to accept the user input
     return {
-      initialValue: exValue,
+      initialValue: value,
       params: result,
     }
   },
 
   binding: {
     reader(_args) {
-      return (exValue: unknown): number => {
+      return (exValue: unknown): string => {
         // Convert an external unknown value into the internal value
-        return typeof exValue === "number" ? exValue : 0
+        return String(exValue)
       }
     },
 
     constraint(args) {
       // Create a value constraint from the user input
-      const constraints = []
-      // You can reuse existing functions of the default plugins
-      const cr = createRangeConstraint(args.params)
-      if (cr) {
-        constraints.push(cr)
-      }
-      const cs = createStepConstraint(args.params)
-      if (cs) {
-        constraints.push(cs)
-      }
+      const constraints: Constraint<string>[] = []
       // Use `CompositeConstraint` to combine multiple constraints
       return new CompositeConstraint(constraints)
     },
@@ -102,12 +83,16 @@ export const TemplateInputPlugin: InputBindingPlugin<
 
   controller(args) {
     // Create a controller for the plugin
-    return new PluginController(args.document, {
+    return new FxStringInputController(args.document, {
       value: args.value,
+      props: ValueMap.fromObject({
+        ...args.params,
+      }),
       viewProps: args.viewProps,
+      parser: (v) => v,
     })
   },
 }
 
 // Export your plugin(s) as constant `plugins`
-export const plugins = [TemplateInputPlugin]
+export const plugins = [FxStringInputPlugin]
