+++ src/components/Params/tweakpane/plugins/StringInputPlugin/style.module.scss
