--- src/context/tweakpane/plugins/StringInputPlugin/style.module.scss
@@ -1,61 +1,0 @@
-// Import core styles
-@import '@tweakpane/core/lib/sass/plugin';
-
-// Additional style for the plugin
-.root {
-	// Extend a general input view style
-	@extend %input;
-
-	cursor: pointer;
-	display: grid;
-	grid-template-columns: repeat(10, 1fr);
-	grid-template-rows: repeat(auto-fit, 10px);
-	height: calc(var(--bld-us) * 3);
-	overflow: hidden;
-	position: relative;
-
-	&.#{$prefix}-v-disabled {
-		opacity: 0.5;
-	}
-
-	.text {
-		// You can use CSS variables for styling. See declarations for details:
-		// ../../node_modules/@tweakpane/core/lib/sass/common/_defs.scss
-		color: var(--in-fg);
-
-		bottom: 2px;
-		font-size: 0.9em;
-		line-height: 0.9;
-		opacity: 0.5;
-		position: absolute;
-		right: 2px;
-	}
-	.dot {
-		height: 10px;
-		line-height: 10px;
-		position: relative;
-		text-align: center;
-
-		&::before {
-			background-color: var(--in-fg);
-			content: '';
-			border-radius: 1px;
-			height: 2px;
-			left: 50%;
-			margin-left: -1px;
-			margin-top: -1px;
-			position: absolute;
-			top: 50%;
-			width: 2px;
-		}
-	}
-	.frac {
-		background-color: var(--in-fg);
-		border-radius: 1px;
-		height: 2px;
-		left: 50%;
-		margin-top: -1px;
-		position: absolute;
-		top: 50%;
-	}
-}
